Prevent double-rendering of content in reviews pages
Removed the unneeded "body-outer" block from the reviews base.html since
it was causing all content in reviews pages to render twice (with the
second time being horribly ugly and sometimes missing data). This was
due to the pyohio-website site_base.html having a "body" block immediately
followed by a "body-outer" block. The reviews templates redefine
"body", which was inserted into the site_base's "body" (as desired), but
having a "body-outer" in reviews that ALSO contained a "body" block
caused the reviews "body" to be placed into the page a second time
(along with the "sidebar" block but let's not even mention that--oh,
drat, too late...).

Anyway. It was confusing to figure out but it's better now.